+++ R/run_blast.R
@@ -15,7 +15,7 @@
 #' @param blast_type One of the available BLAST+ search engines,
 #' #'   one of: `c("blastn", "blastp", "blastx", "tblastn", "tblastx")`.
 #' @param verbose Should condathis::run() internal command be shown?
-#' @param env_name The name of the conda environment with the parameter (i.e. "blast-env")
+#' @param env_name The name of the conda environment with the parameter (i.e. "blastr-blast-env")
 #' @return Unformatted BLAST results.
 #'   For results formatted as tibble, please use `BLASTr::get_blast_results()`
 #'
@@ -33,36 +33,16 @@
 #' }
 #'
 #' @export
-run_blast <- function(asv,
-                      db_path,
-                      num_alignments = 4L,
-                      num_threads = 1L,
-                      blast_type = "blastn",
-                      # gapopen = 5,
-                      # gapextend = 2,
-                      # task = "blastn",
-                      perc_id = 80L,
-                      perc_qcov_hsp = 80L,
-                      verbose = FALSE,
-                      env_name = "blast-env") {
-  #   if (is.null(db_path)) {
-  #   db_path <- getOption(
-  #     "BLASTr.db_path",
-  #     default = NULL
-  #   )
-  # }
-  # if (is.null(db_path)) {
-  #   cli::cli_abort(
-  #     message = "No BLAST database provided."
-  #   )
-  # }
-  # if (is.null(num_threads)) {
-  #   num_threads <- getOption(
-  #     "BLASTr.num_threads",
-  #     default = 1
-  #   )
-  # }
-
+run_blast <- function(
+    asv,
+    db_path,
+    num_alignments = 4L,
+    num_threads = 1L,
+    blast_type = "blastn",
+    perc_id = 80L,
+    perc_qcov_hsp = 80L,
+    verbose = FALSE,
+    env_name = "blastr-blast-env") {
   rlang::check_required(asv)
   rlang::check_required(db_path)
   check_cmd(blast_type, env_name = env_name, verbose = verbose)
@@ -75,9 +55,6 @@ run_blast <- function(asv,
     "-db", db_path,
     "-query", query_path,
     "-outfmt", "6 std qcovhsp staxid",
-    # "-gapopen", gapopen,
-    # "-gapextend", gapextend,
-    # "-task", task,
     "-max_hsps", "1",
     "-perc_identity", perc_id,
     "-qcov_hsp_perc", perc_qcov_hsp,
@@ -91,7 +68,7 @@ run_blast <- function(asv,
   if (isTRUE(blast_res$status != 0L)) {
     error_msg_list <- stringr::str_extract_all(blast_res$stderr, "Error:.*")
 
-    if (length(unlist(error_msg_list)) > 0) {
+    if (isTRUE(length(unlist(error_msg_list)) > 0)) {
       error_msg_vector <- unlist(error_msg_list)
       names(error_msg_vector) <- c(rep("x", times = length(error_msg_vector)))
     } else {